requirements.lock.txt
accelerate==1.10.1
certifi==2026.4.22
charset-normalizer==3.4.7
chronos-forecasting==1.5.3
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
fonttools==4.60.2
fsspec==2025.10.0
hf-xet==1.4.3
huggingface_hub==0.36.2
idna==3.13
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
packaging==26.2
pandas==2.3.3
pillow==11.3.0
psutil==7.2.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytz==2026.2
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.22.2
torch==2.2.2
tqdm==4.67.3
transformers==4.57.6
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
zipp==3.23.1
